requirements.txt
streamlit
requests
Pillow
langdetect
deep-translator
PyMuPDF
python-docx
docx2txt
gTTS
